Divide General settings & Policies pages
